statistics: use float rate (#1838)

+++ server/api/hot_status.go
@@ -27,10 +27,10 @@ type hotStatusHandler struct {
 
 // HotStoreStats is used to record the status of hot stores.
 type HotStoreStats struct {
-	BytesWriteStats map[uint64]uint64 `json:"bytes-write-rate,omitempty"`
-	BytesReadStats  map[uint64]uint64 `json:"bytes-read-rate,omitempty"`
-	KeysWriteStats  map[uint64]uint64 `json:"keys-write-rate,omitempty"`
-	KeysReadStats   map[uint64]uint64 `json:"keys-read-rate,omitempty"`
+	BytesWriteStats map[uint64]float64 `json:"bytes-write-rate,omitempty"`
+	BytesReadStats  map[uint64]float64 `json:"bytes-read-rate,omitempty"`
+	KeysWriteStats  map[uint64]float64 `json:"keys-write-rate,omitempty"`
+	KeysReadStats   map[uint64]float64 `json:"keys-read-rate,omitempty"`
 }
 
 func newHotStatusHandler(handler *server.Handler, rd *render.Render) *hotStatusHandler {
